+++ math/operation.c
@@ -1,11 +1,11 @@
 #include "operation.h"
 
-float add(float a, float b)
+double add(double a, double b)
 {
     return a+b;
 }
 
-float sub(float a, float b)
+double sub(double a, double b)
 {
     return a-b;
 }
