+++ tests/object_test.c
@@ -4,7 +4,7 @@
 
 int main()
 {
-    float vertices[][3] = {{1,2,3},{4,5,6}};
+    double vertices[][3] = {{1,2,3},{4,5,6}};
     int edge[][2] = {{0,1}, {2,3}};
 
     object_t* cube = initialiseObject(vertices, 2, edge, 2);
